Fix web push service worker URL for HTTPS sites
Modify the service worker URL for HTTPS sites to use the existing
service worker from OneSignal's web push WordPress plugin instead of the
service worker bundled with this AMP plugin.

It's slightly better for AMP web push's service worker URL to point to
the existing registered service worker for a subscriber instead of a new
service worker.

  - The service worker already exists in OneSignal's web push plugin.
    Users are enabling a OneSignal-specific AMP web push feature and
    must input OneSignal-specific information like "app ID" and
    "subdomain", so they most likely have OneSignal's web push plugin
    installed.

  - The existing service worker was hardcoded to a specific version of
    OneSignal's web SDK.

  - For users who had customized OneSignal's service worker to merge in
    their own service worker, those users would have to copy their
    customizations to another service worker for AMP pages.

It would be better to provide a configurable service worker URL option,
but this should cover the majority of the cases.